21.2. Listening to An Event From Another Controller
An Ajax Endpoint for Partial HTML
What we really want to do in this method is make an Ajax call to an endpoint that will return the new HTML for the entire cart area. We can create that endpoint with some clever organization.
Start in the template. Copy all of the HTML that's inside of our cart area. So everything that's inside of the cart-list controller element: this div... all the way down to the end. Yep, that looks right.
Now create a new template in templates/cart/ called, how about _cartList.html.twig, and paste.
Back in the original template, include that with {{ include('cart/_cartList.html.twig') }}.
This won't change anything yet: our page still works like it did before.
But now we can add a route & controller that returns just this template partial.
Open src/Controller/CartController.php. This is the controller that renders the shopping cart page. Right below that method, add another one: public function _shoppingCartList().
I'm keeping with the convention of prefixing my template partials - or even controllers that return a "fragment" of HTML - with an underscore. Above this, add @Route() and set the URL to be /cart - to match what's above - and then /_list. Name the route _app_cart_list.
Beautiful! To render the new template, we need one variable: cart... which we get via this CartStorage service that's custom to our project. Copy that argument, paste it down here, and return $this->render('cart/_cartList.html.twig') passing a cart variable set to $cartStorage->getOrCreateCart().
We're done! Go try it in the browser by going directly here. So /cart/_list and... got it! Hit back.
Next: let's update our controller to make an Ajax call to this endpoint and replace the entire cart area with fresh HTML. After we do that, will we need to, somehow, re-initialize our Stimulus controllers on the new HTML elements?
And, as a bonus, we'll add a basic CSS transition to really make things shine.