23.1. Symfony UX & Chart.js -- composer require symfony/ux-chartjs
We now know Stimulus pretty well, which is really just a nice JavaScript library that has nothing to do with Symfony. So then... what exactly is Symfony UX?
To answer that, go to https://github.com/symfony/ux. At its most basic, Symfony UX is a growing list of pre-built Stimulus controllers that you can install to get free functionality. For example, one of them is a controller that integrates Chart.js: a completely independent, lovely JavaScript-powered chart library.
This is perfect... because our sales are really starting to take off. People love our junk, uh, minimalist products! So let's install and use this UX package to build a sales graph on an admin page. I've already created that page at "/admin"... but it's not very interesting yet... and none of these links go anywhere.

Installing symfony/ux-chartjs
Ok: let's check out the docs for Symfony UX chartjs. Scroll down... Whoa. Stimulus and Chart.js are both pure JavaScript libraries. So it's a little interesting that the first step is to install a PHP package. Why are we doing that? Let's find out! Copy the composer require line and head over to your terminal.
I committed all my changes before hitting record so I can easily see any changes that the Flex recipe for this package might make. Paste and go:

composer require symfony/ux-chartjs
While this loads, your Encore build may start failing: don't worry about that for now. When it finishes, run:

git status
Interesting! This modified composer.json and composer.lock, which is totally normal. And the PHP package we just installed is actually a Symfony bundle, so its Flex recipe auto-enabled that bundle in config/bundles.php and updated the symfony.lock file.

UX PHP Packages Come with a JavaScript Package
But here's where things get interesting, really interesting. The recipe also updated our package.json file as well as some other file called assets/controllers.json.

Let's see what changed in package.json:

git diff package.json
Woh! It added a new package! But instead of a version number on the right, it's pointing to a directory inside of vendor/. Let's go check out that path. It was vendor/symfony/ux-chartjs/ then Resources/assets. Cool! This directory is a JavaScript package! How can I tell? It has a package.json file like all JavaScript libraries and, in the src/ directory, a Stimulus controller!
Well, in reality, when we use this package, the compiled controller in the dist/ directory is what will truly be used... but reading the one in src/ is easier.
So: what does this mean for us? It means that each Symfony UX library - like ux-chartjs - is actually two things: a PHP bundle and a JavaScript package. There is also some magic related to how this controller is registered in our app... but we'll talk about that in a few minutes.
Anyways, we have a new Symfony bundle in our vendor/ directory and a new JavaScript package registered in package.json - let me open that - which points to a directory in that bundle.