26.3.React (or Vue) ❤️ Stimulus
Adding the React Component
Now we're ready. Here's our mission: at the bottom of any page, we have a little "made with love" footer. It's a simple example, but I want to rebuild this in React and make it so that if we click the heart, it multiplies.
If you downloaded the course code from this page, you should have a tutorial/ directory with a MadeWithLove.js React component inside. Copy that. Go to assets/. Let's create a new directory called components/ and paste the new file there.
This simple component renders the text, manages a hearts state and increases that heart state on click.

Adding a Stimulus Controller
So... how do we render React components if we're using Stimulus? With a custom Stimulus controller of course!
In assets/controllers/, add a new file called made-with-love_controller.js. Start the same as always... by cheating! Steal the code from counter_controller... add a connect() method and console.log(), this time, a heart.
Next open up templates/base.html.twig and scroll down to find the footer... here it is. Let's replace all of this with our controller. I'll keep the footer <div>... break it onto multiple lines and then add stimulus_controller('made-with-love').
Let's make sure that's connected. Oh, before I do that, I think I forgot to restart Encore:

yarn watch
Awesome. Head over, refresh the page and... perfect. There's our heart!