26.4.React (or Vue) ❤️ Stimulus
Rendering a React Component in Stimulus
With this simple setup, we can now render the React component into our element. First, we need to import a few things: we need ReactDOM from react-dom, react from react and then our component: import MadeWithLove from '../components/MadeWithLove'.
Down in connect, say ReactDOM.render(), pass this our component - <MadeWithLove /> - and then this.element.
That's it! Go back to the page and try it. Yes! That text is coming from our React component! And if we click the heart, it multiplies!
So... yeah! Stimulus and frontend frameworks... are... oddly enough... kind of BFF's.
Want to see something cooler? I'll inspect element on the footer and... hit "Edit as HTML". Right above this, I'll hack in another <div data-controller="made-with-love">... oh, make sure that's data-controller.
When I click off to add that... boom! We get a second, independent React component! Stimulus notices the new element and instantiates a new controller, which, of course, renders a second React app. With this setup, no matter how or when an element gets onto the page, Stimulus will handle booting up the react app.
But stimulus can do even more for React or Vue: it can help us pass data directly from our server into a component as props. It can even do this for full objects! Let's see how next.